apps/mcp-server-own/requirements.txt
# Pinned requirement ranges (conservative). Adjust as needed for your environment.

# Pinned exact versions for reproducible builds (adjust if needed)
fastapi==0.95.2
uvicorn[standard]==0.22.0
fastmcp==0.1.0
requests==2.31.0
psutil==5.9.5
vertica-python==0.1.10
psycopg2-binary==2.9.6
